migrate last react files to typescript
