+++ airflow/www/static/js/api/useGridData.ts
@@ -36,7 +36,7 @@ const dagId = getMetaValue(DAG_ID_PARAM);
 const gridDataUrl = getMetaValue('grid_data_url');
 const urlRoot = getMetaValue('root');
 
-interface GridData {
+export interface GridData {
   dagRuns: DagRun[];
   groups: Task;
   ordering: RunOrdering;
